Refactored to move to user folder and removed additional lines
